Each window has its own details window (#42)
Open different details windows when clicking on the same player on different windows, as they might have different exclusion settings and the data displayed would be different.

Before the data would be added to the same window, causing weird quirks, duplicating the child elements. This was not noticed before barrier and pretty much none had 2 healing stats windows open at the same time but it's easier for any player to spot as one might have a barrier-only window and a healing-only window.